RIA-7173 Refactor of redundant DET notification for ADA suitability
