+++ src/functionalTest/resources/scenarios/RIA-6781-RIA-7173-internal-ada-case-ada-suitability-notification-unsuitable.json
@@ -72,16 +72,12 @@
         "listCaseHearingDate": "2023-05-01T13:30:15.000",
         "notificationsSent":[
           {
-            "id":"6781_ADA_SUITABILITY_DETERMINED_DET",
-            "value":"$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
+            "id": "6781_ADA_SUITABILITY_DETERMINED_INTERNAL_ADA_DET",
+            "value": "$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
           },
           {
             "id":"6781_ADA_SUITABILITY_DETERMINED_HOME_OFFICE",
             "value":"$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
-          },
-          {
-            "id": "6781_ADA_SUITABILITY_DETERMINED_INTERNAL_ADA_DET",
-            "value": "$/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/"
           }
         ]
       }
