--- src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.java
@@ -1,94 +1,0 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.detentionengagementteam;
-
-import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
-
-import com.google.common.collect.ImmutableMap;
-import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
-import lombok.extern.slf4j.Slf4j;
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.RequiredFieldMissingException;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.*;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.IdValue;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.EmailWithLinkNotificationPersonalisation;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DetEmailService;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.DocumentDownloadClient;
-import uk.gov.service.notify.NotificationClientException;
-
-@Slf4j
-@Service
-public class DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation implements EmailWithLinkNotificationPersonalisation {
-
-    private final String internalAdaSuitabilityReviewTemplateId;
-    private final DocumentDownloadClient documentDownloadClient;
-    private final DetEmailService detEmailService;
-
-    public DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation(
-        @Value("${govnotify.template.adaSuitabilityReview.detentionEngagementTeam.internalAda.email}") String internalAdaSuitabilityReviewTemplateId,
-        DetEmailService detEmailService,
-        DocumentDownloadClient documentDownloadClient
-    ) {
-        this.internalAdaSuitabilityReviewTemplateId = internalAdaSuitabilityReviewTemplateId;
-        this.detEmailService = detEmailService;
-        this.documentDownloadClient = documentDownloadClient;
-    }
-
-    @Override
-    public String getTemplateId() {
-        return internalAdaSuitabilityReviewTemplateId;
-    }
-
-    @Override
-    public Set<String> getRecipientsList(AsylumCase asylumCase) {
-        Optional<String> detentionFacility = asylumCase.read(DETENTION_FACILITY, String.class);
-        if (detentionFacility.isEmpty() || !detentionFacility.get().equals("immigrationRemovalCentre")) {
-            return Collections.emptySet();
-        }
-
-        return Collections.singleton(detEmailService.getDetEmailAddress(asylumCase));
-    }
-
-    @Override
-    public String getReferenceId(Long caseId) {
-        return caseId + "_ADA_SUITABILITY_DETERMINED_INTERNAL_ADA_DET";
-    }
-
-    @Override
-    public Map<String, Object> getPersonalisationForLink(AsylumCase asylumCase) {
-        requireNonNull(asylumCase, "asylumCase must not be null");
-
-        return ImmutableMap
-            .<String, Object>builder()
-            .put("appealReferenceNumber", asylumCase.read(AsylumCaseDefinition.APPEAL_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("homeOfficeReferenceNumber", asylumCase.read(AsylumCaseDefinition.HOME_OFFICE_REFERENCE_NUMBER, String.class).orElse(""))
-            .put("appellantGivenNames", asylumCase.read(AsylumCaseDefinition.APPELLANT_GIVEN_NAMES, String.class).orElse(""))
-            .put("appellantFamilyName", asylumCase.read(AsylumCaseDefinition.APPELLANT_FAMILY_NAME, String.class).orElse(""))
-            .put("documentLink", getInternalAdaSuitabilityDocumentInJsonObject(asylumCase))
-            .build();
-    }
-
-    private JSONObject getInternalAdaSuitabilityDocumentInJsonObject(AsylumCase asylumCase) {
-        Optional<List<IdValue<DocumentWithMetadata>>> maybeDocuments = asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS);
-        List<DocumentWithMetadata> documents = maybeDocuments
-            .orElse(Collections.emptyList())
-            .stream()
-            .map(IdValue::getValue)
-            .filter(document -> document.getTag() == DocumentTag.INTERNAL_ADA_SUITABILITY)
-            .collect(Collectors.toList());
-
-        if (documents.size() == 0) {
-            throw new RequiredFieldMissingException("Internal ADA Suitability document is not present");
-        }
-        DocumentWithMetadata document = documents.get(0);
-        try {
-            return documentDownloadClient.getJsonObjectFromDocument(document);
-        } catch (IOException | NotificationClientException e) {
-            log.error("Failed to get Internal ADA Suitability document in compatible format", e);
-            throw new IllegalStateException("Failed to get Internal ADA Suitability document in compatible format");
-        }
-    }
-}
