+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/config/NotificationGeneratorConfiguration.java
@@ -4016,19 +4016,4 @@ public class NotificationGeneratorConfiguration {
                 )
         );
     }
-
-    @Bean("internalAdaSuitabilityReviewNotificationGenerator")
-    public List<NotificationGenerator> internalAdaSuitabilityReviewNotificationGenerator(
-        DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation,
-        GovNotifyNotificationSender notificationSender,
-        NotificationIdAppender notificationIdAppender) {
-
-        return Arrays.asList(
-            new EmailWithLinkNotificationGenerator(
-                newArrayList(Collections.singleton(detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation)),
-                notificationSender,
-                notificationIdAppender
-            )
-        );
-    }
 }
