--- src/test/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/detentionengagementteam/DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisationTest.java
@@ -1,134 +1,0 @@
-package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.detentionengagementteam;
-
-import static com.google.common.collect.Lists.newArrayList;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.TestUtils.*;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
-
-import com.google.common.collect.ImmutableMap;
-import java.io.IOException;
-import java.util.*;
-import org.json.JSONObject;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.junit.jupiter.MockitoSettings;
-import org.mockito.quality.Strictness;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.RequiredFieldMissingException;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentTag;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.DocumentWithMetadata;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.IdValue;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.DetEmailService;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.DocumentDownloadClient;
-import uk.gov.service.notify.NotificationClientException;
-
-@ExtendWith(MockitoExtension.class)
-@MockitoSettings(strictness = Strictness.LENIENT)
-public class DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisationTest {
-
-    @Mock
-    AsylumCase asylumCase;
-    @Mock
-    DetEmailService detEmailService;
-    @Mock
-    JSONObject jsonDocument;
-    @Mock
-    DocumentDownloadClient documentDownloadClient;
-
-    private final String templateId = "someTemplateId";
-    private final String internalAdaSuitabilityReviewPersonalisationReferenceId = "_ADA_SUITABILITY_DETERMINED_INTERNAL_ADA_DET";
-    private final String detEmailAddress = "some@example.com";
-    private final String appealReferenceNumber = "someReferenceNumber";
-    private final String homeOfficeReferenceNumber = "1234-1234-1234-1234";
-    private final String appellantGivenNames = "someAppellantGivenNames";
-    private final String appellantFamilyName = "someAppellantFamilyName";
-    DocumentWithMetadata internalAdaSuitabilityLetter = getDocumentWithMetadata(
-        "1", "ADA-Appellant-letter-suitability-decision-suitable", "some other desc", DocumentTag.INTERNAL_ADA_SUITABILITY);
-    IdValue<DocumentWithMetadata> internalAdaSuitabilityLetterId = new IdValue<>("1", internalAdaSuitabilityLetter);
-    private DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation;
-
-
-    @BeforeEach
-    public void setUp() throws NotificationClientException, IOException {
-
-        when(asylumCase.read(APPEAL_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(appealReferenceNumber));
-        when(asylumCase.read(APPELLANT_GIVEN_NAMES, String.class)).thenReturn(Optional.of(appellantGivenNames));
-        when(asylumCase.read(APPELLANT_FAMILY_NAME, String.class)).thenReturn(Optional.of(appellantFamilyName));
-        when(asylumCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class)).thenReturn(Optional.of(homeOfficeReferenceNumber));
-
-        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.of(newArrayList(internalAdaSuitabilityLetterId)));
-        when(documentDownloadClient.getJsonObjectFromDocument(internalAdaSuitabilityLetter)).thenReturn(jsonDocument);
-
-        detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation =
-            new DetentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation(
-                templateId,
-                detEmailService,
-                documentDownloadClient
-            );
-    }
-
-    @Test
-    public void should_return_given_template_id() {
-        assertEquals(templateId, detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getTemplateId());
-    }
-
-    @Test
-    public void should_return_given_reference_id() {
-        Long caseId = 12345L;
-        assertEquals(caseId + internalAdaSuitabilityReviewPersonalisationReferenceId,
-            detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getReferenceId(caseId));
-    }
-
-    @Test
-    public void should_return_given_email_address_from_asylum_case() {
-        when(asylumCase.read(DETENTION_FACILITY, String.class)).thenReturn(Optional.of("immigrationRemovalCentre"));
-        when(detEmailService.getDetEmailAddress(asylumCase)).thenReturn(detEmailAddress);
-
-        assertTrue(
-            detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getRecipientsList(asylumCase).contains(detEmailAddress));
-    }
-
-    @Test
-    public void should_throw_exception_on_personalisation_when_case_is_null() {
-
-        assertThatThrownBy(
-            () -> detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getPersonalisationForLink((AsylumCase) null))
-            .isExactlyInstanceOf(NullPointerException.class)
-            .hasMessage("asylumCase must not be null");
-    }
-
-    @Test
-    public void should_throw_exception_on_personalisation_when_internal_ada_suitability_review_document_is_missing() {
-        when(asylumCase.read(NOTIFICATION_ATTACHMENT_DOCUMENTS)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(
-            () -> detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getPersonalisationForLink(asylumCase))
-            .isExactlyInstanceOf(RequiredFieldMissingException.class)
-            .hasMessage("Internal ADA Suitability document is not present");
-    }
-
-    @Test
-    public void should_return_personalisation_when_all_information_given_refused() {
-
-        final Map<String, Object> expectedPersonalisation =
-            ImmutableMap
-                .<String, Object>builder()
-                .put("appealReferenceNumber", appealReferenceNumber)
-                .put("homeOfficeReferenceNumber", homeOfficeReferenceNumber)
-                .put("appellantGivenNames", appellantGivenNames)
-                .put("appellantFamilyName", appellantFamilyName)
-                .put("documentLink", jsonDocument)
-                .build();
-
-        Map<String, Object> actualPersonalisation =
-            detentionEngagementTeamInternalAdaSuitabilityReviewPersonalisation.getPersonalisationForLink(asylumCase);
-
-        assertTrue(compareStringsAndJsonObjects(expectedPersonalisation, actualPersonalisation));
-    }
-}
